# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "2.5.10",
    "title": "META API",
    "description": "API文件"
  },
  "host": "six-northnode.onrender.com",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "jwt token"
    }
  },
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/api/movie/": {
      "post": {
        "tags": [
          "Movie"
        ],
        "summary": "新增電影資訊",
        "description": "新增電影資訊",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "type": "object",
            "required": "true",
            "description": "會員註冊用",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "鋼鐵人"
                },
                "$imgs": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "level": {
                  "type": "number",
                  "example": 1
                },
                "desc": {
                  "type": "string",
                  "example": "一個神隱少女的故事"
                },
                "time": {
                  "type": "number",
                  "example": 180
                },
                "actors": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "videos": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "status": {
                  "type": "number",
                  "example": 1
                },
                "releaseData": {
                  "type": "string",
                  "example": "2023-05-01"
                }
              },
              "required": [
                "name",
                "level",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "回傳範例資料",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "捍衛任務4"
                },
                "imgs": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "level": {
                  "type": "number",
                  "example": 1
                },
                "desc": {
                  "type": "string",
                  "example": "一個神隱少女的故事"
                },
                "time": {
                  "type": "number",
                  "example": 180
                },
                "actors": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "videos": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "status": {
                  "type": "number",
                  "example": 1
                },
                "releaseData": {
                  "type": "string",
                  "example": "2023-05-01T12:00:00Z"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Movie"
        ],
        "summary": "獲取電影列表",
        "description": "獲取電影列表",
        "parameters": [
          {
            "name": "isRelease\r",
            "in": "query"
          },
          {
            "name": "name",
            "in": "query",
            "description": "電影名稱模糊搜尋",
            "type": "string",
            "default": "鋼鐵人"
          },
          {
            "name": "isRelease",
            "description": "是否上檔",
            "type": "boolean",
            "default": true,
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "回傳範例資料",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number",
                        "example": 1
                      },
                      "name": {
                        "type": "string",
                        "example": "鋼鐵人"
                      },
                      "imgs": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "level": {
                        "type": "number",
                        "example": 0
                      },
                      "desc": {
                        "type": "string",
                        "example": "string"
                      },
                      "time": {
                        "type": "number",
                        "example": 90
                      },
                      "actors": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "videos": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "status": {
                        "type": "number",
                        "example": 1
                      },
                      "release_data": {
                        "type": "string",
                        "example": "yymmdd-hms"
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  },
...
}